# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/agent/git-token": {
            "get": {
                "security": [
                    {
                        "AgentAuth": []
                    }
                ],
                "description": "Returns a GitHub installation token for repository access. Rate limited.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent"
                ],
                "summary": "Request Git token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg_server.AgentGitTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_pullbase_pullbase_server_pkg_apierrors.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_pullbase_pullbase_server_pkg_apierrors.Error"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/github_com_pullbase_pullbase_server_pkg_apierrors.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_pullbase_pullbase_server_pkg_apierrors.Error"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/github_com_pullbase_pullbase_server_pkg_apierrors.Error"
                        }
                    }
                }
            }
        },
        "/agent/server-info": {
            "get": {
                "security": [
                    {
                        "AgentAuth": []
                    }
                ],
                "description": "Returns repository configuration for a server. Authenticated via agent token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent"
                ],
                "summary": "Get server info (Agent Token)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg_server.ServerInfoResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_pullbase_pullbase_server_pkg_apierrors.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_pullbase_pullbase_server_pkg_apierrors.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_pullbase_pullbase_server_pkg_apierrors.Error"
                        }
                    }
                }
            }
        },
        "/agent/status": {
            "post": {
                "security": [
                    {
                        "AgentAuth": []
                    }
                ],
                "description": "Receives status updates from agents including drift detection and errors. Authenticated via agent token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent"
                ],
                "summary": "Update agent status (Agent Token)",
                "parameters": [
                    {
                        "description": "Status update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_server.AgentStatusPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_pullbase_pullbase_server_pkg_apierrors.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_pullbase_pullbase_server_pkg_apierrors.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_pullbase_pullbase_server_pkg_apierrors.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_pullbase_pullbase_server_pkg_apierrors.Error"
                        }
                    }
                }
            }
        },
...
        "/servers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a paginated list of servers with their latest status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "List all servers",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum number of servers to return (1-500)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Number of servers to skip",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field (name, created_at, status)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_pullbase_pullbase_server_pkg_models.ServerWithStatus"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg_server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pkg_server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg_server.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new server with an initial agent token for deployment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servers"
                ],
                "summary": "Create a new server",
                "parameters": [
                    {
                        "description": "Server creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_server.CreateServerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/pkg_server.CreateServerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg_server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pkg_server.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/pkg_server.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/pkg_server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg_server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_pullbase_pullbase_server_pkg_apierrors.Error": {
            "type": "object",
            "properties": {
                "err": {
                    "description": "Underlying error, if any"
                },
                "kind": {
                    "description": "The category of error",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_pullbase_pullbase_server_pkg_apierrors.Kind"
                        }
                    ]
                },
                "message": {
                    "description": "Human-readable error message",
                    "type": "string"
                },
                "op": {
                    "description": "Operation that failed (e.g., \"GetServer\", \"CreateUser\")",
                    "type": "string"
                }
            }
        },
...
        "pkg_server.AgentGitTokenResponse": {
            "type": "object",
            "properties": {
                "app_slug": {
                    "type": "string"
                },
                "authentication": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "installation_id": {
                    "type": "integer"
                },
                "provider": {
                    "$ref": "#/definitions/github_com_pullbase_pullbase_server_pkg_models.GitProvider"
                },
                "repo_url": {
                    "type": "string"
                },
                "repository_id": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "pkg_server.AgentStatusPayload": {
            "type": "object",
            "properties": {
                "agent_version": {
                    "type": "string"
                },
                "commit_hash": {
                    "type": "string"
                },
                "drift_details": {
                    "$ref": "#/definitions/github_com_pullbase_pullbase_server_pkg_models.DriftDetails"
                },
                "error_message": {
                    "type": "string"
                },
                "is_drifted": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "pkg_server.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "pkg_server.ServerInfoResponse": {
            "type": "object",
            "properties": {
                "auto_reconcile": {
                    "type": "boolean"
                },
                "branch": {
                    "type": "string"
                },
                "deploy_path": {
                    "type": "string"
                },
                "repo_url": {
                    "type": "string"
                },
                "target_commit_hash": {
                    "type": "string"
                }
            }
        },
...
    }
}